requirements.txt
arrow
python-dotenv
pandas
requests
numpy
matplotlib
ipython
musicbrainzngs
tekore
aiohttp
